refactor sending to inbox out of post POST method
